refactor: move FabricStorageUtil to storage package
