+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/storage/FabricStorageInsertableStorage.java
@@ -4,7 +4,6 @@ import com.refinedmods.refinedstorage2.api.core.Action;
 import com.refinedmods.refinedstorage2.api.storage.Actor;
 import com.refinedmods.refinedstorage2.api.storage.InsertableStorage;
 import com.refinedmods.refinedstorage2.platform.api.exporter.AmountOverride;
-import com.refinedmods.refinedstorage2.platform.fabric.util.FabricStorageUtil;
 
 import java.util.function.Function;
 
@@ -16,18 +15,18 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerLevel;
 
-public class StorageInsertableStorage<T, P> implements InsertableStorage<T> {
+public class FabricStorageInsertableStorage<T, P> implements InsertableStorage<T> {
     private final BlockApiCache<Storage<P>, Direction> cache;
     private final Function<T, P> toPlatformMapper;
     private final Direction direction;
     private final AmountOverride amountOverride;
 
-    public StorageInsertableStorage(final BlockApiLookup<Storage<P>, Direction> lookup,
-                                    final Function<T, P> toPlatformMapper,
-                                    final ServerLevel serverLevel,
-                                    final BlockPos pos,
-                                    final Direction direction,
-                                    final AmountOverride amountOverride) {
+    public FabricStorageInsertableStorage(final BlockApiLookup<Storage<P>, Direction> lookup,
+                                          final Function<T, P> toPlatformMapper,
+                                          final ServerLevel serverLevel,
+                                          final BlockPos pos,
+                                          final Direction direction,
+                                          final AmountOverride amountOverride) {
         this.cache = BlockApiCache.create(lookup, serverLevel, pos);
         this.toPlatformMapper = toPlatformMapper;
         this.direction = direction;
